// types.ts
export interface OptimizationTask {
  fileName: string; // The clean output filename (e.g. "folder/image.png")
  relativePath: string; // The lookup path for reference
  originalWidth: number;
  originalHeight: number;
  targetWidth: number;
  targetHeight: number;
  blob: Blob; // The in-memory source data
  maxScaleUsed: number;
  isResize: boolean;
  overridePercentage?: number;
}